combines axios and fetch into one interface and enforces interception
