--- axios.d.ts
@@ -1,3 +1,0 @@
-import SuperTokens from './lib/build/axios';
-
-export default SuperTokens;
